Fix passing 64bit non-float parameters on GPRs.
When storing a 64bit parameter, an array of instructions to store in GPRs is
used. To use the array properly r19 should all the time contain base address
of the array.

The code for storing first word of a 64bit non-float parameter was updating
r19 so on the store phase for the second word of 64bit parameter the array
base address was incorrect.

The patch makes sure the array base address in r19 remains intact.